% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_clean.roclet_tests.R
\name{roclet_clean.roclet_tests}
\alias{roclet_clean.roclet_tests}
\title{Delete Unit Test Files Generated by @tests Tags}
\usage{
\method{roclet_clean}{roclet_tests}(x, base_path)
}
\arguments{
\item{x}{A `roclet` object.}

\item{base_path}{Path to root of source package.}
}
\value{
`NULL`  If found, files created by `roxut` will be deleted.
}
\description{
Delete Unit Test Files Generated by @tests Tags
}
\author{
Bryan A. Hanson
}
