% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_output.roclet_tests.R
\name{roclet_output.roclet_tests}
\alias{roclet_output.roclet_tests}
\title{Roclet to Write the Results of Processing @tests Tags to Unit Test File}
\usage{
\method{roclet_output}{roclet_tests}(x, results, base_path, ...)
}
\arguments{
\item{x}{A `roclet` object.}

\item{results}{Value returned from the `roclet_process()` method.}

\item{base_path}{Path to root of source package.}

\item{...}{Other arguments passed downstream. Needed for compatibility,
does not appear to be used.}
}
\value{
`NULL`
}
\description{
Roclet to Write the Results of Processing @tests Tags to Unit Test File
}
\author{
Bryan A. Hanson
}
