% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_process.roclet_tests.R
\name{roclet_process.roclet_tests}
\alias{roclet_process.roclet_tests}
\title{Process @tests Tags}
\usage{
\method{roclet_process}{roclet_tests}(x, blocks, env, base_path)
}
\arguments{
\item{x}{A `roclet` object.}

\item{blocks}{A list of [roxy_block] objects.}

\item{env}{Package environment.}

\item{base_path}{Path to root of source package.}
}
\value{
A list containing XXX.
}
\description{
Process @tests Tags
}
\author{
Bryan A. Hanson
}
