% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxut-package.R
\docType{package}
\name{roxut-package}
\alias{roxut-package}
\title{roxut}
\description{
Much as `roxygen2` allows one to document functions in the same file
as the function itself, `roxut` allows one to write the unit tests
in the same file as the function.  Once processed, the unit tests are
moved to the appropriate directory.  Currently supports `testthat` and
`tinytest` frameworks.  The `roxygen2` package provides much of the infrastructure.
}
\author{
Bryan A. Hanson.

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{package}
