% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxut.R
\name{roxut}
\alias{roxut}
\title{Process a package using the tests roclet}
\usage{
roxut(package.dir = ".", load_code = NULL, clean = FALSE)
}
\arguments{
\item{package.dir}{Location of package top level directory. Default is
working directory.}

\item{load_code}{A function used to load all the R code in the package
directory. The default, `NULL`, uses the strategy defined by
the `load` `roxygen2` option, which defaults to [load_pkgload()].
See [load] for more details.}

\item{clean}{If `TRUE`, `roxygen2` will delete all files previously
created by `roxut` before running the tests roclet.  If the file
names have not changed existing files are overwritten each time
`roxut` is run.}
}
\value{
`NULL`  Files will be written to the appropriate directory.
}
\description{
This is the user-facing function that converts `@tests` tags
written in the `*.R` function definition files into unit test files.
This function follows `roxygenize()` closely and is adapted from it.
}
\details{
This function works by examining any files in the `R` subdirectory of
a package, looking for any `@tests` tags which start a block of unit tests.
Either the `tinytest` or `testthat` framework can be used.  Each unit test
is copied to the appropriate directory, with one file per `*R` file.
It is up to the package author to
create and deploy the supervising functions required by each framework.
}
\author{
Bryan A. Hanson, adapted and modified from `roxygen2::roxygenize()`
}
