% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy_tag_parse.roxy_tag_tests.R
\name{roxy_tag_parse.roxy_tag_tests}
\alias{roxy_tag_parse.roxy_tag_tests}
\title{Parse @tests Tags}
\usage{
\method{roxy_tag_parse}{roxy_tag_tests}(x)
}
\arguments{
\item{x}{A `roclet` object.}
}
\value{
The `x` object is returned, appended with a list `val`
       containing two elements:
\itemize{
  \item `framework` containing the name of the unit test framework in use.
  \item `tests`containing the content of the unit tests.
}
}
\description{
The `roxygen2` infrastructure is used to parse the `@tests` tags.
This function is called once for each tag in a `*.R` file.
}
\author{
Bryan A. Hanson
}
