% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_roclet.R
\name{tests_roclet}
\alias{tests_roclet}
\title{tests_roclet}
\usage{
tests_roclet()
}
\value{
XXX
}
\description{
This roclet automates the creation of unit tests, using `@tests` tags
written in the `*.R` function definition files, and processed using `roxygen2`.
The unit tests are written verbatim to the appropriate destination file.
}
\author{
Bryan A. Hanson
}
