% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_output.roclet_tests.R
\name{roclet_output.roclet_tests}
\alias{roclet_output.roclet_tests}
\title{Write the Results of Processing @tests Tags to Unit Test File}
\usage{
\method{roclet_output}{roclet_tests}(x, results, base_path, ...)
}
\arguments{
\item{x}{A \code{roclet} object.}

\item{results}{Value returned from the \code{roclet_process()} method.}

\item{base_path}{Path to root of source package.}

\item{...}{Other arguments passed downstream. Needed for compatibility,
does not appear to be used.}
}
\value{
Invisibly, a list with the text of the tests. Writes files!
}
\description{
The contents of the unit tests (lines following the \code{@tests} tags)
are written verbatim to the appropriate directories (directory depends
on whether \code{tinytest} or \code{testthat} is in use).  If there is more
than one function and corresponding unit tests in a file, the first unit tests
will be in \code{filename.R}, the second in \code{filename-1.R} etc.
}
\author{
Bryan A. Hanson
}
