% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy_tag_parse.roxy_tag_tests.R
\name{roxy_tag_parse.roxy_tag_tests}
\alias{roxy_tag_parse.roxy_tag_tests}
\title{Parse @tests Tags}
\usage{
\method{roxy_tag_parse}{roxy_tag_tests}(x)
}
\arguments{
\item{x}{A \code{roclet} object.}
}
\value{
The \code{x} object is returned, appended with a list \code{val}
       containing two elements:
\itemize{
  \item \code{framework} containing the name of the unit test framework in use.
  \item \code{tests} containing the content of the unit tests.
}
}
\description{
This function is called once for each \code{@tests} tag in a
function definition file.
}
\author{
Bryan A. Hanson
}
