% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_roclet.R
\name{tests_roclet}
\alias{tests_roclet}
\title{tests_roclet}
\usage{
tests_roclet()
}
\value{
A complete roclet.
}
\description{
This roclet automates the creation of unit tests, using \code{@tests} tags
written in the function definition files, and processed using \code{roxygen2}.
The unit tests are written verbatim to the appropriate destination file.
}
\author{
Bryan A. Hanson
}
